+++ packages/expo/test/index.test.js
@@ -1,30 +1,34 @@
+
 const delivery = require('@bugsnag/delivery-expo')
 
 jest.mock('expo-constants', () => ({
   default: {
     platform: {},
-    expoConfig: {}
+    expoConfig: {},
+    expoGoConfig: null
   }
 }))
 
-jest.mock('../../plugin-expo-device/node_modules/expo-constants', () => ({
+jest.mock('../../../node_modules/expo-constants', () => ({
   default: {
     platform: {},
-    expoConfig: {}
+    expoConfig: {},
+    expoGoConfig: null
   }
 }))
 
-jest.mock('../../plugin-expo-app/node_modules/expo-application', () => ({}))
+jest.mock('../../../node_modules/expo-application', () => ({}))
 
-jest.mock('../../plugin-expo-app/node_modules/expo-constants', () => ({
+jest.mock('../../../node_modules/expo-constants', () => ({
   default: {
     platform: {},
-    expoConfig: {}
+    expoConfig: {},
+    expoGoConfig: null
   }
 }))
 
 jest.mock('@bugsnag/delivery-expo')
-jest.mock('../../delivery-expo/node_modules/expo-crypto', () => ({}))
+jest.mock('../../../node_modules/expo-crypto', () => ({}))
 
 jest.mock('react-native', () => ({
   NativeModules: {
@@ -54,7 +58,7 @@ jest.mock('react-native', () => ({
   }
 }))
 
-jest.mock('../../delivery-expo/node_modules/expo-file-system', () => ({
+jest.mock('../../../node_modules/expo-file-system', () => ({
   cacheDirectory: 'file://var/data/foo.bar.app/',
   downloadAsync: jest.fn(() => Promise.resolve({ md5: 'md5', uri: 'uri' })),
   getInfoAsync: jest.fn(() => Promise.resolve({ exists: true, md5: 'md5', uri: 'uri' })),
@@ -68,20 +72,34 @@ jest.mock('../../delivery-expo/node_modules/expo-file-system', () => ({
   createDownloadResumable: jest.fn(() => Promise.resolve())
 }))
 
-jest.mock('../../delivery-expo/node_modules/@react-native-community/netinfo', () => ({
+jest.mock('../../../node_modules/@react-native-community/netinfo', () => ({
   addEventListener: jest.fn(),
   fetch: () => Promise.resolve({ isConnected: true })
 }))
 
-jest.mock('../../plugin-expo-connectivity-breadcrumbs/node_modules/@react-native-community/netinfo', () => ({
+jest.mock('../../../node_modules/@react-native-community/netinfo', () => ({
   addEventListener: jest.fn(),
   fetch: () => Promise.resolve({ isConnected: true })
 }))
 
-jest.doMock('../../plugin-expo-device/node_modules/expo-device', () => ({
+jest.doMock('../../../node_modules/expo-device', () => ({
   manufacturer: 'Google',
   modelName: 'Pixel 4'
 }))
+
+jest.doMock('expo-secure-store', () => ({
+  getItem: () => 'c0123456789abcdef0123456789',
+  setItem: () => {},
+  ALWAYS_THIS_DEVICE_ONLY: 4
+}))
+
+const networkBreadcrumbsPlugin = {
+  load: () => () => {}
+}
+
+jest.mock('../../../node_modules/@bugsnag/plugin-network-breadcrumbs', () => {
+  return () => networkBreadcrumbsPlugin
+})
 
 global.ErrorUtils = {
   setGlobalHandler: jest.fn(),
@@ -228,6 +246,61 @@ describe('expo notifier', () => {
       }), expect.any(Function))
 
       done()
+    })
+  })
+
+  describe('configuration', () => {
+    beforeEach(() => {
+      jest.resetModules()
+    })
+
+    it('sets a default value for releaseStage correctly (production)', () => {
+      jest.mock('expo-constants', () => ({
+        default: {
+          platform: {},
+          expoConfig: {},
+          expoGoConfig: null
+        }
+      }))
+
+      const config = require('../src/config')
+      expect(config.releaseStage.defaultValue()).toBe('production')
+    })
+
+    it('sets a default value for releaseStage correctly (local-dev)', () => {
+      jest.mock('expo-constants', () => ({
+        default: {
+          platform: {},
+          expoConfig: null,
+          expoGoConfig: {
+            developer: {
+              tool: 'expo-cli'
+            }
+          }
+        }
+      }))
+
+      global.__DEV__ = true
+      const config = require('../src/config')
+      expect(config.releaseStage.defaultValue()).toBe('local-dev')
+    })
+
+    it('sets a default value for releaseStage correctly (local-prod)', () => {
+      jest.mock('expo-constants', () => ({
+        default: {
+          platform: {},
+          expoConfig: {
+            developer: {
+              tool: 'expo-cli'
+            }
+          },
+          expoGoConfig: null
+        }
+      }))
+
+      global.__DEV__ = false
+      const config = require('../src/config')
+      expect(config.releaseStage.defaultValue()).toBe('local-prod')
     })
   })
 })