+++ CHANGELOG.md
@@ -1,6 +1,53 @@
 # Changelog
 
-## v48.1.0 (2023-03-27)
+## [51.0.1] - 2024-08-29
+
+### Fixed
+
+- (plugin-expo-device) Do not use expo-secure-store on unsupported platforms [#185](https://github.com/bugsnag/bugsnag-expo/pull/185)
+- (plugin-expo-device, plugin-expo-app) Replace Constants.platform usage with Platform API [#185](https://github.com/bugsnag/bugsnag-expo/pull/185)
+
+## [51.0.0] - 2024-05-16
+
+This release adds support for Expo 51.
+
+### Breaking Changes
+
+This release contains an additional peer dependency on `expo-secure-store`. See the [upgrade guide](UPGRADING.md) for more information.
+
+### Changed
+
+- (plugin-expo-device) Replace `Constants.installationId` with a generated device id [#181](https://github.com/bugsnag/bugsnag-expo/pull/181)
+
+## [50.0.0] - 2024-02-07
+
+This release adds support for expo 50
+
+## [49.0.2] - 2023-12-07
+
+### Added
+
+- (delivery-expo) Explicitly mark failed payloads >1MB as not retryable [#65](https://github.com/bugsnag/bugsnag-expo/pull/65)
+
+### Fixed
+
+- (plugin-expo-eas-sourcemaps) Ensure EAS sourcemap config plugin is idempotent [#156](https://github.com/bugsnag/bugsnag-expo/pull/156)
+
+## [49.0.1] - 2023-08-03
+
+### Fixed
+
+- Fix a crash in configuration when Constants.expoGoConfig is null [#145](https://github.com/bugsnag/bugsnag-expo/pull/145)
+
+## [49.0.0] - 2023-07-13
+
+This release adds support for expo 49
+
+### Changed
+
+- Replace deprecated Constants.manifest [#141](https://github.com/bugsnag/bugsnag-expo/pull/141)
+
+## [48.1.0] - 2023-03-27
 
 ### Added
 
@@ -12,7 +59,7 @@
 - (plugin-expo-eas-sourcemaps) Reinstate API key in Android manifest [#117](https://github.com/bugsnag/bugsnag-expo/pull/117)
 - (plugin-expo-eas-sourcemaps) Support dynamic configuration files in EAS Build lifecycle hook [#117](https://github.com/bugsnag/bugsnag-expo/pull/117)
 
-## v48.0.0 (2023-03-07)
+## [48.0.0] - 2023-03-07
 
 This release adds support for expo 48
 
@@ -21,13 +68,13 @@ This release adds support for expo 48
 - (bugsnag-expo-cli) CLI tool now installs a sourcemap plugin version that matches the Expo SDK version [#111](https://github.com/bugsnag/bugsnag-expo/pull/111)
 - (plugin-expo-eas-sourcemaps) Use EAS Build lifecycle hook for Android source map uploads [#112](https://github.com/bugsnag/bugsnag-expo/pull/112)
 
-## v47.1.1 (2023-03-02)
+## [47.1.1] - 2023-03-02
 
 ### Fixed
 
 - (plugin-expo-eas-sourcemaps) Restrict Bugsnag Android Gradle Plugin dependency to v7 [#104](https://github.com/bugsnag/bugsnag-expo/pull/104)
 
-## v47.1.0 (2023-01-09)
+## [47.1.0] - 2023-01-09
 
 ### Fixed
 
@@ -38,34 +85,34 @@ This release adds support for expo 48
 - Added `app.versionCode` (Android) and `app.bundleVersion` (iOS) metadata for standalone apps [#92](https://github.com/bugsnag/bugsnag-expo/pull/92)
 - Added `codeBundleId` configuration option [#92](https://github.com/bugsnag/bugsnag-expo/pull/92)
 
-## v47.0.0 (2022-11-21)
+## [47.0.0] - 2022-11-21
 
 This release adds support for expo 47
 
-## v46.0.2 (2022-11-21)
+## [46.0.2] - 2022-11-21
 
 ### Fixed
 
 - (bugsnag-expo-cli) Fix issue with automated installation when using app.config.js [#71](https://github.com/bugsnag/bugsnag-expo/pull/71)
 
-## v46.0.1 (2022-09-22)
+## [46.0.1] - 2022-09-22
 
 ### Fixed
 
 - (bugsnag-expo-cli) Improve monorepo compatibility for plugin-expo-eas-sourcemaps installation [#59](https://github.com/bugsnag/bugsnag-expo/pull/59)
 
-## v46.0.0 (2022-09-09)
+## [46.0.0] - 2022-09-09
 
 This release adds support for expo 46
 
-## v45.1.1 (2022-08-04)
+## [45.1.1] - 2022-08-04
 
 ### Added
 
 - Added `Bugsnag.isStarted()` to check whether Bugsnag has initialized [#34](https://github.com/bugsnag/bugsnag-expo/pull/34)
 - (plugin-expo-eas-sourcemaps) Add minimum version check to sourcemap plugin [#45](https://github.com/bugsnag/bugsnag-expo/pull/45)
 
-## v45.1.0 (2022-07-28)
+## [45.1.0] - 2022-07-28
 
 This release adds support for sourcemap uploads when using EAS Build
 
@@ -74,17 +121,17 @@ This release adds support for sourcemap uploads when using EAS Build
 - Added new `@bugsnag/plugin-expo-eas-sourcemaps` plugin [#40](https://github.com/bugsnag/bugsnag-expo/pull/40)
 - Updated CLI tool to automate installation of sourcemap plugin [#40](https://github.com/bugsnag/bugsnag-expo/pull/40)
 
-## v45.0.0 (2022-05-23)
+## [45.0.0] - 2022-05-23
 
 This release adds support for expo 45
 
-## v44.0.1 (2022-05-12)
+## [44.0.1] - 2022-05-12
 
-### Fixes
+### Fixed
 
 - Fixed an issue in the `bugsnag-expo-cli` install script when reporting an unsupported version of expo [#31](https://github.com/bugsnag/bugsnag-expo/pull/31)
 
-## v44.0.0 (2022-04-19)
+## [44.0.0] - 2022-04-19
 
 This release marks a change in the version scheme used by `@bugsnag/expo` and a move to its own repo, [`bugsnag-expo`](https://github.com/bugsnag/bugsnag-expo)